refractor: fixed pr review bugs
