+++ src/pages/Components/CardsList.js/CardsList.module.css
@@ -1,7 +1,3 @@
-body {
-    color: #1e1e1e;
-}
-
 .categ-list {
     display: flex;
     align-items: flex-start;
