+++ src/pages/Components/Navigation/Navigation.css
@@ -1,7 +1,3 @@
-body {
-    font-family:  Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-}
-
 .nav-header {
     background-color: #ffffff80;
     font-size: 1.5rem;
