+++ src/pages/Components/SlideShow/SlideShow.module.css
@@ -1,4 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;900&display=swap');
 
 .slideShow {
     position: relative;
