+++ src/pages/Components/index.jsx
@@ -5,3 +5,5 @@ export { GridDisplay } from "./GridCardDisplay/GridDisplay"
 export { Footer } from "./Footer/Footer"
 
 export { CardsList } from "./CardsList.js/CardsList"
+
+export { SlideShow } from "./SlideShow/SlideShow"
