+++ src/pages/index.jsx
@@ -4,3 +4,6 @@ export { Navigation, Footer } from "./Components/";
 export { Cart } from "./Cart/Cart"
 export { SignIn } from "./SignIn/SignIn"
 export { SignUp } from "./SignUp/SignUp"
+export { Searchbar } from "./Searchbar/Searchbar"
+export { Wishlist } from "./Wishlist/Wishlist"
+export { PageNotFound } from "./PageNotFound/PageNotFound"
